backend/requirements.txt
beautifulsoup4==4.12.2
fastapi==0.104.0
langchain==0.0.321
nltk==3.8.1
pinecone_client==2.2.4
pydantic==1.10.9
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
Requests==2.31.0
torch==2.0.1
uvicorn==0.23.2
transformers==4.31.0
sentence-transformers==2.2.2
datasets==2.14.0
accelerate==0.21.0
einops==0.6.1
bitsandbytes==0.41.0
pypdf==3.16.1